+++ Specs/Widgets/CesiumWidget/CesiumWidgetSpec.js
@@ -175,33 +175,25 @@ defineSuite([
         }).toThrowDeveloperError();
     });
 
-    it('raises renderLoopError and stops the render loop when render throws', function() {
+    it('stops the render loop when render throws', function() {
         widget = new CesiumWidget(container);
         expect(widget.useDefaultRenderLoop).toEqual(true);
 
-        var spyListener = jasmine.createSpy('listener');
-        widget.renderLoopError.addEventListener(spyListener);
-
         var error = 'foo';
-        widget.render = function() {
+        widget.scene.primitives.update = function() {
             throw error;
         };
 
         waitsFor(function() {
-            return spyListener.wasCalled;
-        });
-
-        runs(function() {
-            expect(spyListener).toHaveBeenCalledWith(widget, error);
-            expect(widget.useDefaultRenderLoop).toEqual(false);
-        });
+            return !widget.useDefaultRenderLoop;
+        }, 'render loop to be disabled.');
     });
 
     it('shows the error panel when render throws', function() {
         widget = new CesiumWidget(container);
 
         var error = 'foo';
-        widget.render = function() {
+        widget.scene.primitives.update = function() {
             throw error;
         };
 
@@ -226,7 +218,7 @@ defineSuite([
         });
 
         var error = 'foo';
-        widget.render = function() {
+        widget.scene.primitives.update = function() {
             throw error;
         };
 
@@ -237,37 +229,5 @@ defineSuite([
         runs(function() {
             expect(widget._element.querySelector('.cesium-widget-errorPanel')).toBeNull();
         });
-    });
-
-    it('raises the preRender event prior to rendering', function() {
-        widget = new CesiumWidget(container);
-
-        var preRenderInvocations = 0;
-        widget.preRender.addEventListener(function() {
-            ++preRenderInvocations;
-        });
-
-        widget.resize();
-        widget.render();
-
-        expect(preRenderInvocations).toBe(1);
-
-        widget.destroy();
-    });
-
-    it('raises the postRender event after rendering', function() {
-        widget = new CesiumWidget(container);
-
-        var postRenderInvocations = 0;
-        widget.postRender.addEventListener(function() {
-            ++postRenderInvocations;
-        });
-
-        widget.resize();
-        widget.render();
-
-        expect(postRenderInvocations).toBe(1);
-
-        widget.destroy();
     });
 }, 'WebGL');