+++ CHANGES.md
@@ -7,9 +7,9 @@ Beta Releases
 ### b29 - 2014-06-02
 
 * Breaking changes
-  * Renamed `CesiumWidget.onRenderLoopError` to `CesiumWidget.renderLoopError`.
+  * Removed `CesiumWidget.onRenderLoopError` and `Viewer.renderLoopError`.  They have been replaced by `Scene.renderError`.
 * Improved terrain and imagery rendering performance when very close to the surface.
-* Added `preRender` and `postRender` events to `CesiumWidget` and `Viewer`.
+* Added `preRender` and `postRender` events to `Scene`.
 
 ### b28 - 2014-05-01